move core folder to root
