+++ source-code/sdk-js/src/adapter-sveltekit/runtime/client/reactive/index.ts
@@ -1,7 +1,7 @@
 import type * as Ast from "@inlang/core/ast"
 import { getContext, setContext } from "svelte"
 import { writable, type Readable } from "svelte/store"
-import type { RelativeUrl } from "../../../../core/index.js"
+import type { RelativeUrl } from "../../../../index.js"
 import { inlangSymbol } from "../../shared/utils.js"
 import type { SvelteKitClientRuntime } from "../index.js"
 import type * as Runtime from "../../../../runtime/index.js"
